pyproject.toml
[tool.poetry]
name = "algebraic_nas_string_kernels"
version = "0.1.0"
description = ""
authors = [
    "Lum Birinxhiku <birinxhl@cs.uni-freiburg.de>",
    "Simon Schrodi <schrodi@cs.uni-freiburg.de>",
    "Danny Stoll <stolld@cs.uni-freiburg.de>",
]
readme = "README.md"
license = ""
packages = [
    { include = "skanas", from = "src" },
]

[tool.poetry.dependencies]
# we need to limit the python versions allowed until the `skanas` branch of `neps`
# can use newer versions of python
# TODO: fix when #skanas_upgraded
python = ">=3.7.1,<3.8"
hydra-core = "^1.1.1"
# in the main branch of `neps` the package is called `neural-pipeline-search`
# however, we use the old name `neps` since `skanas` branch still uses it
# TODO: fix when #skanas_upgraded
neps = { path = "../neps", develop = true }
python-git-info = "^0.7.1"
PyYAML = "^6.0"
torch = "^1.11.0+cu102"
numpy = "^1.21.4"
pandas = "^1.3.5"
scikit-learn = "0.24.2"
seaborn = "^0.11.2"
matplotlib = "^3.5.1"
torchmetrics = "^0.8.1"
torchvision = "^0.13.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
pytest-cov = "^3"
jupyter = "^1.0"
isort = "^5.10.1"
black = "^22.8.0"
mypy = "^0.971"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
minversion = "3.7"
addopts = "-vv"

[tool.coverage.run]
branch = true
context = "skanas"
omit = []

[tool.coverage.report]
show_missing = true
skip_covered = true
exclude_lines = [
    "pragma: no cover",
    "raise NotImplementedError",
    "if TYPE_CHECKING",
]

[tool.black]
target-version = ['py37']

[tool.isort]
profile = 'black'
src_paths = ["skanas"]
lines_after_imports = 2
known_first_party = ["skanas", "neps"]

[tool.mypy]
files = ["src/skanas"]
namespace_packages = true
follow_imports = "skip"
ignore_missing_imports = true
show_error_codes = true
incremental = false
explicit_package_bases = true

[tool.pylint.messages_control]
disable = 'all'
enable = [
    'invalid-characters-in-docstring','wrong-spelling-in-comment','wrong-spelling-in-docstring','not-in-loop',
    'function-redefined','continue-in-finally','abstract-class-instantiated','star-needs-assignment-target',
    'duplicate-argument-name','return-in-init','too-many-star-expressions','nonlocal-and-global',
    'return-outside-function','return-arg-in-generator','invalid-star-assignment-target','bad-reversed-sequence',
    'nonexistent-operator','yield-outside-function','init-is-generator','nonlocal-without-binding','lost-exception',
    'assert-on-tuple','dangerous-default-value','duplicate-key','useless-else-on-loop','expression-not-assigned',
    'confusing-with-statement','unnecessary-lambda','pointless-statement','unnecessary-pass','unreachable','eval-used',
    'exec-used','using-constant-test','deprecated-lambda','blacklisted-name','misplaced-comparison-constant',
    'singleton-comparison','unneeded-not','ßconsider-iterating-dictionary','consider-using-enumerate','empty-docstring',
    'unidiomatic-typecheck','condition-evals-to-constant','consider-using-generator','nan-comparison',
    'consider-using-min-max-builtin','consider-using-with','invalid-all-format','consider-using-dict-items',
    'deprecated-decorator','forgotten-debug-statement','unspecified-encoding','useless-with-lock',
    'use-implicit-booleaness-not-comparison','not-async-context-manager','yield-inside-async-function',
    'await-outside-async','invalid-unary-operand-type','unsupported-binary-operation','not-callable',
    'redundant-keyword-arg','assignment-from-no-return','assignment-from-none','not-context-manager',
    'repeated-keyword','missing-kwoa','no-value-for-parameter','invalid-sequence-index','invalid-slice-index',
    'unexpected-keyword-arg','unsupported-membership-test','unsubscriptable-object','bad-except-order',
    'catching-non-exception','bad-exception-context','notimplemented-raised','raising-bad-type','raising-non-exception',
    'misplaced-bare-raise','duplicate-except','nonstandard-exception','binary-op-exception','bare-except',
    'raise-missing-from','consider-using-namedtuple-or-dataclass','consider-using-tuple','bad-open-mode',
    'redundant-unittest-assert','boolean-datetime','deprecated-methodimport-error','import-self','reimported',
    'relative-import','deprecated-module','wildcard-import','misplaced-future','cyclic-import','wrong-import-position',
    'ungrouped-imports','multiple-imports','simplifiable-condition','len-as-condition',
    'unpacking-non-sequence','invalid-all-object','unbalanced-tuple-unpacking','undefined-variable',
    'undefined-all-variable','used-before-assignment','cell-var-from-loop','global-variable-undefined',
    'redefined-builtin','redefine-in-handler','unused-import','unused-argument','unused-wildcard-import',
    'unused-variable','global-variable-not-assigned','undefined-loop-variable','global-statement',
    'global-at-module-level','format-needs-mapping','truncated-format-string','missing-format-string-key',
    'mixed-format-string','too-few-format-args','bad-str-strip-call','too-many-format-args','bad-format-character',
    'format-combined-specification','bad-format-string-key','bad-format-string','missing-format-attribute',
    'missing-format-argument-key','unused-format-string-argument','unused-format-string-key','invalid-format-index',
    'f-string-without-interpolation','use-maxsplit-arg','anomalous-unicode-escape-in-string',
    'anomalous-backslash-in-string','redundant-u-string-prefix','format-string-without-interpolation',
    'simplifiable-if-statement','logging-format-truncated','logging-too-few-args','logging-too-many-args',
    'logging-unsupported-format','not-an-iterable','not-a-mapping','use-sequence-for-iteration','bad-indentation',
    'unnecessary-semicolon','missing-final-newline','mixed-line-endings','multiple-statements','trailing-newlines',
    'trailing-whitespace','unexpected-line-ending-format','superfluous-parens','access-member-before-definition',
    'method-hidden','assigning-non-slot','duplicate-bases','inconsistent-mro','inherit-non-class','invalid-slots',
    'invalid-slots-object','no-method-argument','no-self-argument','unexpected-special-method-signature',
    'non-iterator-returned','invalid-length-returned','protected-access','attribute-defined-outside-init',
    'abstract-method','bad-staticmethod-argument','non-parent-init-called','super-init-not-called',
    'no-classmethod-decorator','no-staticmethod-decorator','no-self-use','bad-classmethod-argument',
    'bad-mcs-classmethod-argument','bad-mcs-method-argument','method-check-failed','invalid-bool-returned',
    'invalid-index-returned','invalid-repr-returned','invalid-str-returned','invalid-bytes-returned',
    'invalid-hash-returned','invalid-length-hint-returned','invalid-format-returned','invalid-getnewargs-returned',
    'invalid-getnewargs-ex-returned','super-with-arguments','deprecated-class','invalid-class-object',
    'unused-private-member',
]
